Add XXH3 checksum support
Add XXH3-64 and XXH3-128 hashing through xxHash v0.8.3, including file filters, save formats, verification parsing and options UI support.

Update property/verify tests and CI vector setup and refresh NIST vector download URLs.

+++ HashProp.c
@@ -125,7 +125,7 @@ VOID __fastcall HashPropWorkerMain( PHASHPROPCONTEXT phpctx )
 
 	// Which checksum types we want to calculate
     // (this is loaded earlier in HashPropDlgInit())
-	DWORD checksumFlags = (UINT8)phpctx->opt.dwChecksums;
+	DWORD checksumFlags = phpctx->opt.dwChecksums;
 
 	phpctx->dwReadBufferSize = 0;
 	phpctx->bOuterMultithreaded = FALSE;
